backend/requirements.txt
fastapi==0.115.5
pydantic== 2.10.2
motor==3.6.0
python-dotenv==1.0.1
uvicorn==0.32.1
